# federation/src/main/java/com/exam/federation/dto/CollectivityInformation.java
package com.exam.federation.dto;

import lombok.Data;

@Data
public class CollectivityInformation {
    private String id;
    private String number;
    private String name;
    private String location;
}