# federation/src/main/java/com/exam/federation/dto/CreateActivityMemberAttendance.java
package com.exam.federation.dto;

import com.exam.federation.entity.Enums.AttendanceStatus;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.AllArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class CreateActivityMemberAttendance {
    private String memberIdentifier;
    private AttendanceStatus attendanceStatus;
}